refactor(api): modularize GraphQL schema and loader structure
- Split GraphQL schema into multiple files (enums, types, queries, mutations, subscriptions)
- Added schema loader utility to merge .graphql files dynamically
- Centralized schema creation logic in a dedicated module
- Updated directory structure for better maintainability
