+++ apps/api/package.json
@@ -34,7 +34,9 @@
     "pg": "^8.16.3",
     "prisma": "^6.13.0",
     "ws": "^8.18.3",
-    "zod": "^4.0.14"
+    "zod": "^4.0.14",
+    "@graphql-tools/load-files": "^7.0.1",
+    "@graphql-tools/merge": "^9.1.1"
   },
   "devDependencies": {
     "@types/body-parser": "^1.19.6",
